# docs/swagger/swagger.yaml
basePath: /v1
definitions:
  dto.AuthResponse:
    properties:
      tenant_id:
        type: string
      token:
        type: string
      user_id:
        type: string
    type: object
...
  dto.CreateCustomerRequest:
    properties:
      address_city:
        maxLength: 100
        type: string
      address_country:
        type: string
      address_line1:
        maxLength: 255
        type: string
      address_line2:
        maxLength: 255
        type: string
      address_postal_code:
        maxLength: 20
        type: string
      address_state:
        maxLength: 100
        type: string
      email:
        type: string
      external_id:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
    required:
    - external_id
    type: object
...
  dto.CustomerResponse:
    properties:
      address_city:
        description: AddressCity is the city of the customer's address
        type: string
      address_country:
        description: AddressCountry is the country of the customer's address (ISO
          3166-1 alpha-2)
        type: string
      address_line1:
        description: AddressLine1 is the first line of the customer's address
        type: string
      address_line2:
        description: AddressLine2 is the second line of the customer's address
        type: string
      address_postal_code:
        description: AddressPostalCode is the postal code of the customer's address
        type: string
      address_state:
        description: AddressState is the state of the customer's address
        type: string
      created_at:
        type: string
      created_by:
        type: string
      email:
        description: Email is the email of the customer
        type: string
      environment_id:
        description: EnvironmentID is the environment identifier for the customer
        type: string
      external_id:
        description: ExternalID is the external identifier for the customer
        type: string
      id:
        description: ID is the unique identifier for the customer
        type: string
      metadata:
        additionalProperties:
          type: string
        description: Metadata
        type: object
      name:
        description: Name is the name of the customer
        type: string
      status:
        $ref: '#/definitions/types.Status'
      tenant_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  dto.ListCustomersResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.CustomerResponse'
        type: array
      pagination:
        $ref: '#/definitions/types.PaginationResponse'
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - email
    - password
    type: object
...
  dto.SignUpRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      tenant_name:
        type: string
      token:
        type: string
    required:
    - email
    type: object
...
  dto.UpdateCustomerRequest:
    properties:
      address_city:
        maxLength: 100
        type: string
      address_country:
        type: string
      address_line1:
        maxLength: 255
        type: string
      address_line2:
        maxLength: 255
        type: string
      address_postal_code:
        maxLength: 20
        type: string
      address_state:
        maxLength: 100
        type: string
      email:
        type: string
      external_id:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
    type: object
...
  errors.ErrorDetail:
    properties:
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
  errors.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/errors.ErrorDetail'
      success:
        type: boolean
    type: object
...
  types.PaginationResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  types.Status:
    enum:
    - published
    - deleted
    - archived
    type: string
    x-enum-varnames:
    - StatusPublished
    - StatusDeleted
    - StatusArchived
...
info:
  contact: {}
  description: FlexPrice API Service
  title: FlexPrice API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login a user
      parameters:
      - description: Login request
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Login
      tags:
      - Auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Sign up a new user
      parameters:
      - description: Sign up request
        in: body
        name: signup
        required: true
        schema:
          $ref: '#/definitions/dto.SignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Sign up
      tags:
      - Auth
  /customers:
    get:
      consumes:
      - application/json
      description: Get customers
      parameters:
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: customer_ids
        type: array
      - in: query
        name: email
        type: string
      - in: query
        name: end_time
        type: string
      - in: query
        name: expand
        type: string
      - in: query
        name: external_id
        type: string
      - in: query
        maximum: 1000
        minimum: 1
        name: limit
        type: integer
      - in: query
        minimum: 0
        name: offset
        type: integer
      - enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - in: query
        name: sort
        type: string
      - in: query
        name: start_time
        type: string
      - enum:
        - published
        - deleted
        - archived
        in: query
        name: status
        type: string
        x-enum-varnames:
        - StatusPublished
        - StatusDeleted
        - StatusArchived
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListCustomersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get customers
      tags:
      - Customers
    post:
      consumes:
      - application/json
      description: Create a customer
      parameters:
      - description: Customer
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCustomerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a customer
      tags:
      - Customers
  /customers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a customer
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a customer
      tags:
      - Customers
    get:
      consumes:
      - application/json
      description: Get a customer
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a customer
      tags:
      - Customers
    put:
      consumes:
      - application/json
      description: Update a customer
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      - description: Customer
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCustomerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a customer
      tags:
      - Customers
...
  /customers/lookup/{lookup_key}:
    get:
      consumes:
      - application/json
      description: Get a customer by lookup key (external_id)
      parameters:
      - description: Customer Lookup Key (external_id)
        in: path
        name: lookup_key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a customer by lookup key
      tags:
      - Customers
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Enter your API key in the format *x-api-key &lt;api-key&gt;**
    in: header
    name: x-api-key
    type: apiKey
swagger: "2.0"